[cmds] Rename (old) tar.c to tar-sash.c, (new) ntar.c to tar.c
